requirements.txt
setuptools==36.2.7
hvac==0.2.17
pytest==3.2.1
requests==2.18.4
responses==0.7.0
python_daemon==2.1.2
